refactor(types-devices): remove redundant type exports
