+++ libs/device-core/src/capabilities/devicesWithTouchScreen.test.ts
@@ -1,0 +1,12 @@
+import { DeviceModelId } from "@ledgerhq/devices";
+import { DevicesWithTouchScreen } from "./devicesWithTouchScreen";
+
+describe("DevicesWithTouchScreen", () => {
+  it("includes only touchscreen device models", () => {
+    expect(DevicesWithTouchScreen).toEqual([
+      DeviceModelId.stax,
+      DeviceModelId.europa,
+      DeviceModelId.apex,
+    ]);
+  });
+});
