+++ libs/device-core/src/capabilities/devicesWithTouchScreen.ts
@@ -1,0 +1,10 @@
+import { DeviceModelId } from "@ledgerhq/devices";
+
+/**
+ * Device models equipped with a touchscreen.
+ */
+export const DevicesWithTouchScreen = [
+  DeviceModelId.stax,
+  DeviceModelId.europa,
+  DeviceModelId.apex,
+] as const;
