+++ libs/ledger-live-common/src/deviceSDK/hooks/useBatteryStatuses.ts
@@ -1,12 +1,11 @@
 import { useCallback, useEffect, useState } from "react";
 import { log } from "@ledgerhq/logs";
-import { BatteryStatusFlags } from "@ledgerhq/types-devices";
 import {
   getBatteryStatusesAction,
   GetBatteryStatusesActionState,
   initialState,
 } from "../actions/getBatteryStatuses";
-import { BatteryStatusTypes } from "../../hw/getBatteryStatus";
+import { BatteryStatusTypes, type BatteryStatusFlags } from "../../hw/getBatteryStatus";
 import { useEnv } from "@features/platform-env";
 
 export type UseBatteryStatusesArgs = {
