+++ libs/live-dmk-desktop/src/transport/DeviceManagementKitTransport.ts
@@ -5,10 +5,9 @@ import {
   type DeviceSessionState,
   DiscoveredDevice,
 } from "@ledgerhq/device-management-kit";
-import Transport from "@ledgerhq/hw-transport";
+import Transport, { type DescriptorEvent } from "@ledgerhq/hw-transport";
 import { dmkToLedgerDeviceIdMap, activeDeviceSessionSubject } from "@ledgerhq/live-dmk-shared";
 import { LocalTracer } from "@ledgerhq/logs";
-import { DescriptorEvent } from "@ledgerhq/types-devices";
 import { firstValueFrom, Observer, startWith, pairwise, map, Subscription } from "rxjs";
 import { getDeviceManagementKit } from "../hooks/useDeviceManagementKit";
