+++ libs/live-dmk-mobile/src/transport/DeviceManagementKitBLETransport.ts
@@ -1,4 +1,8 @@
-import Transport from "@ledgerhq/hw-transport";
+import Transport, {
+  type DescriptorEvent,
+  type Observer as TransportObserver,
+  type Subscription as TransportSubscription,
+} from "@ledgerhq/hw-transport";
 import {
   DeviceId,
   DeviceManagementKit,
@@ -24,11 +28,7 @@ import {
   Subscription,
 } from "rxjs";
 import { first, filter, tap, timeout, retry, map } from "rxjs/operators";
-import { DescriptorEvent, DeviceModel } from "@ledgerhq/types-devices";
-import type {
-  Observer as TransportObserver,
-  Subscription as TransportSubscription,
-} from "@ledgerhq/hw-transport";
+import type { DeviceModel } from "@ledgerhq/types-devices";
 import { HwTransportError } from "@ledgerhq/hw-transport/errors";
 import { PairingFailed, PeerRemovedPairing, isPeerRemovedPairingError } from "../errors";
 import { getDeviceManagementKit } from "../hooks/useDeviceManagementKit";
